LPD-62844 Use selected items in the info panels as provided by the FDS. This avoids the need for an extra event and so prevents double-rendering of the info panel. In addition, let react manage <AssetTypeInfoPanel> component lifecycle by avoiding direct function call from props transformer.
