+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/info_panel/AssetTypeInfoPanelContent.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 
 import {
 	ISearchAssetObjectEntry,
@@ -15,33 +15,18 @@ import {AssetTypeInfoPanelContext, IAssetTypeInfoPanelContext} from './context';
 
 import '../../../css/components/AssetTypeInfoPanel.scss';
 import {getBaseAssetInformation} from './util';
-import {EVENTS} from './util/constants';
 
-const AssetTypeInfoPanelContent = ({additionalProps: {cmsGroupId}}: any) => {
-	const [assetInfo, setAssetInfo] = useState(
-		{} as ISearchAssetTypeInformation
-	);
-	const [objectEntries, setObjectEntries] = useState(
-		[] as ISearchAssetObjectEntry[]
-	);
-
-	useEffect(() => {
-		const handler = ({items}: {items: ISearchAssetObjectEntry[]}): void => {
-			setObjectEntries(items as ISearchAssetObjectEntry[]);
-		};
-
-		Liferay.on(EVENTS.ASSET_DATA, handler);
-
-		return () => {
-			Liferay.detach(EVENTS.ASSET_DATA, handler);
-		};
-	}, [setObjectEntries]);
-
-	useEffect(() => {
-		if (objectEntries.length === 1) {
-			setAssetInfo(getBaseAssetInformation(objectEntries[0]));
-		}
-	}, [objectEntries]);
+const AssetTypeInfoPanelContent = ({
+	additionalProps: {cmsGroupId},
+	items: objectEntries,
+}: {
+	additionalProps: any;
+	items: ISearchAssetObjectEntry[];
+}) => {
+	const assetInfo: ISearchAssetTypeInformation =
+		objectEntries?.length === 1
+			? getBaseAssetInformation(objectEntries[0])
+			: {};
 
 	return (
 		<>
