+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/AllFDSPropsTransformer.tsx
@@ -4,28 +4,24 @@
  */
 
 import {IInternalRenderer} from '@liferay/frontend-data-set-web';
+import React from 'react';
 
+import {ISearchAssetObjectEntry} from '../../structure_builder/types/AssetType';
 import AssetTypeInfoPanel from '../info_panel/AssetTypeInfoPanelContent';
-import {EVENTS} from '../info_panel/util/constants';
 import createAssetAction from './actions/createAssetAction';
-import createFolderAction from './actions/createFolderAction';
 import multipleFilesUploadAction from './actions/multipleFilesUploadAction';
 import AuthorRenderer from './cell_renderers/AuthorRenderer';
 import NameRenderer from './cell_renderers/NameRenderer';
-import SimpleActionLinkRenderer from './cell_renderers/SimpleActionLinkRenderer';
+import SpaceRenderer from './cell_renderers/SpaceRenderer';
 import TypeRenderer from './cell_renderers/TypeRenderer';
 import addOnClickToCreationMenuItems from './utils/addOnClickToCreationMenuItems';
 
 const ACTIONS = {
 	createAsset: createAssetAction,
-	createFolder: createFolderAction,
 	uploadMultipleFiles: multipleFilesUploadAction,
 };
 
-const OBJECT_ENTRY_FOLDER_CLASSNAME =
-	'com.liferay.object.model.ObjectEntryFolder';
-
-export default function FolderFDSPropsTransformer({
+export default function AllFDSPropsTransformer({
 	creationMenu,
 	itemsActions = [],
 	...otherProps
@@ -56,8 +52,8 @@ export default function FolderFDSPropsTransformer({
 					type: 'internal',
 				} as IInternalRenderer,
 				{
-					component: SimpleActionLinkRenderer,
-					name: 'simpleActionLinkTableCellRenderer',
+					component: SpaceRenderer,
+					name: 'spaceTableCellRenderer',
 					type: 'internal',
 				} as IInternalRenderer,
 				{
@@ -67,7 +63,9 @@ export default function FolderFDSPropsTransformer({
 				} as IInternalRenderer,
 			],
 		},
-		infoPanelComponent: () => AssetTypeInfoPanel(otherProps),
+		infoPanelComponent: (items: {items: ISearchAssetObjectEntry[]}) => (
+			<AssetTypeInfoPanel {...(otherProps as any)} {...items} />
+		),
 		itemsActions: itemsActions.map((action) => {
 			if (action?.data?.id === 'download') {
 				return {
@@ -76,32 +74,8 @@ export default function FolderFDSPropsTransformer({
 						Boolean(item?.embedded?.file?.link?.href),
 				};
 			}
-			else if (action?.data?.id === 'actionLink') {
-				return {
-					...action,
-					isVisible: (item: any) =>
-						Boolean(
-							item?.entryClassName !==
-								OBJECT_ENTRY_FOLDER_CLASSNAME
-						),
-				};
-			}
 
 			return action;
 		}),
-		onActionDropdownItemClick: ({
-			action,
-			itemData,
-		}: {
-			action: any;
-			itemData: [];
-		}) => {
-			if (action?.data?.id === 'show-details') {
-				Liferay.fire(EVENTS.ASSET_DATA, {items: [{...itemData}]});
-			}
-		},
-		onSelectedItemsChange: (selectedItems: any[]) => {
-			Liferay.fire(EVENTS.ASSET_DATA, {items: selectedItems});
-		},
 	};
 }
